learning implemented
learning mechanism added as Gavi required (talked to him to make sure it
is good).

+++ ServerImpl/Entities/UserLevel.cs
@@ -20,8 +20,8 @@ namespace Entities
         [Required]
         public int level { get; set; }
         [Required]
-        public int timesAnswered { get; set; }
+        public double timesAnswered { get; set; }
         [Required]
-        public int timesAnsweredCorrectly { get; set; }
+        public double timesAnsweredCorrectly { get; set; }
     }
 }
